// src/interfaces/Escala.ts
export interface Escala {
   DATA: string;
   FUNÇÃO: string;
   INICIO: string;
   TERMINO: string;
   NOME: string;
   RE: string;
   ID: string;
   DATA2: string;
   FUNÇÃO2: string;
   INICIO2: string;
   TERMINO2: string;
   NOME2: string;
   RE2: string;
   ID2: string;
   DATA3: string;
   FUNÇÃO3: string;
   INICIO3: string;
   TERMINO3: string;
   NOME3: string;
   RE3: string;
   ID3: string;
   DATA4: string;
   FUNÇÃO4: string;
   INICIO4: string;
   TERMINO4: string;
   NOME4: string;
   RE4: string;
   ID4: string;
 }
 